Bearer and gprs error handlers updated
Check the actual bearer status before retry.
If bearer is already open and an ERROR response received, do not
consider this as an error. Add new function to check bearer status.
Update the function activate_gprs() to handle ERROR response.
Update select_network() function to automatically switch to the
'automatic network registration mode' on a failed attempt to register
manually.