Redesign the UART communication using interrupts
The UART reception from the GSM modem is now handled using interrupts.
The interrupt driven code is only implemented for the Zephyr framework
now. A global buffer will store the received bytes, which
can be accessed by other modules. The ISR can be used by
specifying an expected acknowledgement string and a timeout.