Make the UART buffer size configurable
An external buffer can be passed to the library, which can be used to
store the bytes received over the UART. The buffer size may be
configured at compile time (board specific) using a Kconfig in the
calling module. The buffer size depends on the flash page size of
each MCU, as required for OTA binary transfer.