Encapsulate the state of the modem class instead of using globals.
The commit is a bit hard to follow; it started life as trying to make
the RX buffer configurable on the fly, to be able to use smaller or
larger buffers depending on the volume of data expected.

First the `resp_buf` and `resp_buf_len` have been moved inside the
GPRS class. This led to moving all the global state variables as class
members and adapting and polishing the constructor and uart interrupt
handler.

No proper member access control has been done yet.

Notable changes are:

- the ack message buffer has been reduced to a maximum of 12
characters, down from 32 (never used for more than expecting "OK")

- new method available `void set_rx_buf(uint8_t *, size_t)` that
allows consumers to swap the buffer used for the next operation, so as
not to owervrite a buffer that we are currently procesing, e.g while
having to send a short ACK back.